refactor(cli): extract export parser owner
