+++ crates/tokmd/src/cli/parser.rs
@@ -14,8 +14,6 @@
 //! * I/O operations (except config file parsing)
 //! * Higher-tier crate dependencies
 //!
-use std::path::PathBuf;
-
 pub use crate::tool_schema::ToolSchemaFormat;
 use clap::{Args, Parser, Subcommand, ValueEnum};
 use serde::{Deserialize, Serialize};
@@ -27,6 +25,7 @@ mod cockpit;
 mod completions;
 mod context;
 mod diff;
+mod export;
 mod gate;
 mod init;
 mod lang;
@@ -46,6 +45,7 @@ pub use context::{
     CliContextArgs, ContextOutput, ContextStrategy, HandoffArgs, HandoffPreset, ValueMetric,
 };
 pub use diff::{ColorMode, DiffArgs, DiffFormat};
+pub use export::CliExportArgs;
 pub use gate::{CliGateArgs, GateFormat};
 pub use init::{InitArgs, InitProfile};
 pub use lang::CliLangArgs;
@@ -415,53 +415,6 @@ pub enum Commands {
 
     /// Run as a conforming sensor, producing a SensorReport.
     Sensor(SensorArgs),
-}
-
-#[derive(Args, Debug, Clone)]
-pub struct CliExportArgs {
-    /// Paths to scan (directories, files, or globs). Defaults to "."
-    #[arg(value_name = "PATH")]
-    pub paths: Option<Vec<PathBuf>>,
-
-    /// Output format [default: jsonl].
-    #[arg(long, value_enum)]
-    pub format: Option<ExportFormat>,
-
-    /// Write output to this file instead of stdout.
-    #[arg(long, value_name = "PATH", visible_alias = "out")]
-    pub output: Option<PathBuf>,
-
-    /// Module roots (see `tokmd module`) [default: crates,packages].
-    #[arg(long, value_delimiter = ',')]
-    pub module_roots: Option<Vec<String>>,
-
-    /// Module depth (see `tokmd module`) [default: 2].
-    #[arg(long, visible_alias = "depth")]
-    pub module_depth: Option<usize>,
-
-    /// Whether to include embedded languages (tokei "children" / blobs) [default: separate].
-    #[arg(long, value_enum)]
-    pub children: Option<ChildIncludeMode>,
-
-    /// Drop rows with fewer than N code lines [default: 0].
-    #[arg(long)]
-    pub min_code: Option<usize>,
-
-    /// Stop after emitting N rows (0 = unlimited) [default: 0].
-    #[arg(long)]
-    pub max_rows: Option<usize>,
-
-    /// Include a meta record (JSON / JSONL only). Enabled by default.
-    #[arg(long, action = clap::ArgAction::Set)]
-    pub meta: Option<bool>,
-
-    /// Redact paths (and optionally module names) for safer copy/paste into LLMs [default: none].
-    #[arg(long, value_enum)]
-    pub redact: Option<RedactMode>,
-
-    /// Strip this prefix from paths before output (helps when paths are absolute).
-    #[arg(long, value_name = "PATH")]
-    pub strip_prefix: Option<PathBuf>,
 }
 
 // =============================================================================
